refactor: extract status badge into a reusable component
